requirements.txt
alembic==1.13.1
filedepot==0.11.0
Flask==3.0.3
#2.3.3
#Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
#itsdangerous==1.1.0
#Jinja2==2.11.2
#Mako==1.1.4
#MarkupSafe==1.1.1
PyMySQL==1.1.0
passlib==1.7.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
#python-editor==1.0.4
speaklater==1.3
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
sqlalchemy-filters==0.13.0
twilio==9.0.5
Pillow==10.3.0
scrypt==0.8.24
